@@ -9,16 +9,23 @@
         "dotenv": "^8.2.0",
         "express": "^4.17.1",
         "google-auth-library": "^5.5.1",
+        "ts-node": "^8.6.2",
         "uuid": "^3.3.3"
     },
     "devDependencies": {
-        "eslint": "^6.6.0",
+        "@types/express": "^4.17.2",
+        "@types/uuid": "^3.4.7",
+        "@typescript-eslint/eslint-plugin": "^2.18.0",
+        "@typescript-eslint/parser": "^2.18.0",
+        "eslint": "^6.8.0",
         "eslint-config-airbnb-base": "^14.0.0",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-promise": "^4.2.1"
+        "eslint-plugin-import": "^2.20.0",
+        "eslint-plugin-promise": "^4.2.1",
+        "typescript": "^3.7.5"
     },
     "scripts": {
-        "lint": "eslint . --fix",
+        "start": "ts-node app.ts",
+        "lint": "eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "repository": {
